# docs/api.yaml
openapi: 3.0.3
...
info:
  title: HP / VVS Sales Automation Service API
  version: 1.0.0
  description: >-
    HTTP API for the HP / VVS Sales Automation local service. All field names are
    derived from the Context Pack and remain the single source of truth.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /payments/record:
    post:
      summary: Record or update a payments ledger document
      operationId: recordPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRecordRequest'
            example:
              anchorType: SO
              rootApptId: HP-2001
              soNumber: SO-2001
              docType: Deposit Receipt
              paymentDateTime: 2024-07-04T10:15:00-07:00
              method: Card
              reference: AUTH-001
              notes: Initial deposit
              amountGross: 200.0
              feePercent: 2.75
              lines:
                - desc: Deposit
                  qty: 1
                  amt: 200.0
      responses:
        '200':
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRecordResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Payments feature disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payments/summary:
    get:
      summary: Summarize payments for a sales order or appointment
      operationId: summarizePayments
      parameters:
        - in: query
          name: rootApptId
          schema:
            type: string
          description: Filter ledger rows for the given RootApptID.
        - in: query
          name: soNumber
          schema:
            type: string
          description: Filter ledger rows for the given sales order number.
      responses:
        '200':
          description: Payments summary for the requested anchor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentsSummaryResponse'
        '400':
          description: Missing or invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Payments feature disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    PaymentLineItem:
      type: object
      properties:
        desc:
          type: string
          description: Line description.
        qty:
          type: number
          format: double
          description: Quantity applied to the line.
        amt:
          type: number
          format: double
          description: Amount per unit for the line.
        lineTotal:
          type: number
          format: double
          description: Quantity multiplied by amount; populated by the server.
    PaymentRecordRequest:
      type: object
      required:
        - anchorType
        - docType
        - paymentDateTime
        - method
        - amountGross
      properties:
        anchorType:
          type: string
          enum: [SO, APPT]
        rootApptId:
          type: string
          nullable: true
        soNumber:
          type: string
          nullable: true
        docNumber:
          type: string
          nullable: true
          description: Optional explicit ledger document number; otherwise generated.
        docRole:
          type: string
          nullable: true
          description: Optional override for the inferred document role.
        docType:
          type: string
          enum:
            - Deposit Invoice
            - Deposit Receipt
            - Sales Invoice
            - Sales Receipt
            - Credit Memo
            - Payment Receipt
        docStatus:
          type: string
          nullable: true
          enum: [DRAFT, ISSUED]
        paymentDateTime:
          type: string
          format: date-time
        method:
          type: string
          enum: [Card, Wire, Zelle, Cash, Check, Other]
        reference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        amountGross:
          type: number
          format: double
        feePercent:
          type: number
          format: double
          nullable: true
        feeAmount:
          type: number
          format: double
          nullable: true
        lines:
          type: array
          items:
            $ref: '#/components/schemas/PaymentLineItem'
        amountNet:
          type: number
          format: double
          nullable: true
          description: Optional client-provided net amount; the service recomputes the value.
        subtotal:
          type: number
          format: double
          nullable: true
          description: Optional subtotal provided by the client; the service recomputes from lines.
    PaymentRecordResponse:
      type: object
      required:
        - status
        - docNumber
        - docRole
        - docType
        - docStatus
        - paymentDateTime
        - method
        - amountGross
        - amountNet
        - lines
        - requestHash
      properties:
        status:
          type: string
          enum: [CREATED, UPDATED]
        docNumber:
          type: string
        docRole:
          type: string
        anchorType:
          type: string
        rootApptId:
          type: string
          nullable: true
        soNumber:
          type: string
          nullable: true
        docType:
          type: string
        docStatus:
          type: string
        paymentDateTime:
          type: string
          format: date-time
        method:
          type: string
        reference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        amountGross:
          type: number
          format: double
        feePercent:
          type: number
          format: double
          nullable: true
        feeAmount:
          type: number
          format: double
          nullable: true
        subtotal:
          type: number
          format: double
        amountNet:
          type: number
          format: double
        lines:
          type: array
          items:
            $ref: '#/components/schemas/PaymentLineItem'
        requestHash:
          type: string
    PaymentSummaryEntry:
      type: object
      required:
        - docNumber
        - docRole
        - docType
        - docStatus
        - amountGross
        - amountNet
      properties:
        docNumber:
          type: string
        docRole:
          type: string
        anchorType:
          type: string
        rootApptId:
          type: string
          nullable: true
        soNumber:
          type: string
          nullable: true
        docType:
          type: string
        docStatus:
          type: string
        paymentDateTime:
          type: string
          format: date-time
          nullable: true
        method:
          type: string
          nullable: true
        reference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        amountGross:
          type: number
          format: double
        feePercent:
          type: number
          format: double
          nullable: true
        feeAmount:
          type: number
          format: double
          nullable: true
        subtotal:
          type: number
          format: double
          nullable: true
        amountNet:
          type: number
          format: double
        lines:
          type: array
          items:
            $ref: '#/components/schemas/PaymentLineItem'
        requestHash:
          type: string
        submittedAt:
          type: string
          format: date-time
          nullable: true
    PaymentsSummaryResponse:
      type: object
      required:
        - invoicesLinesSubtotal
        - totalPayments
        - netLinesMinusPayments
        - byMethod
        - entries
      properties:
        invoicesLinesSubtotal:
          type: number
          format: double
        totalPayments:
          type: number
          format: double
        netLinesMinusPayments:
          type: number
          format: double
        byMethod:
          type: object
          additionalProperties:
            type: number
            format: double
        entries:
          type: array
          items:
            $ref: '#/components/schemas/PaymentSummaryEntry'
...